requirements.txt
Flask==3.1.1
stripe==12.3.0
python-dotenv==1.0.1
requests==2.32.4
gunicorn